pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-rules-generator"
version = "1.0.0"
description = "A standard for defining project rules and skills for AI agents."
authors = [
  { name = "Amitro123", email = "amitro123@example.com" },
]
readme = "README.md"
requires-python = ">=3.8"
keywords = ["ai", "agents", "rules", "skills", "scaffolding"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pathspec>=0.11.0",
    "pydantic>=2.0.0",
    "tqdm>=4.66.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "groq>=0.4.0"
]

[project.scripts]
project-rules-generator = "main:cli"
prg = "main:cli"

[tool.setuptools]
py-modules = ["main", "project_rules_generator"]
packages = ["generator", "analyzer", "prg_utils", "src"]

[tool.setuptools.package-data]
generator = ["skills/**/*"]
